+++ src/controllers/web/server/Oauth2AccessTokenAction.php
@@ -25,8 +25,7 @@ class Oauth2AccessTokenAction extends Oauth2BaseServerAction
             $psr7Response = $server->respondToAccessTokenRequest($psr7Request, $psr7Response);
             return Psr7Helper::psr7ToYiiResponse($psr7Response);
         } catch (\Exception $e) {
-            Yii::error((string)$e, __METHOD__);
-            return $this->processException($e);
+            return $this->processException($e, __METHOD__);
         }
     }
 }